% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{AssignLeftClust}
\alias{AssignLeftClust}
\title{Assign the left points}
\usage{
AssignLeftClust(overlap.clu, sum.clu, data.simi, n = n)
}
\arguments{
\item{overlap.clu}{The label of each overlap clusters.}

\item{sum.clu}{The number of overlap clusters}

\item{data.simi}{Similarity matrix}

\item{n}{The number of observations}
}
\description{
Assign the left points to the cluster which is of largest similarity
}
\keyword{internal}
