% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{BossaSimi}
\alias{BossaSimi}
\title{Bossa Similarity}
\usage{
BossaSimi(data, is.pca = TRUE, pca.sum.prop = 0.95, fix.pca.comp = FALSE,
  n.comp = 50, alpha = 1)
}
\arguments{
\item{data}{A data.frame or matrix(\code{n*p}) of original categorical data.}

\item{is.pca}{A logical variable indicating if the Bossa scores should transformed
to principle components and then calculate the similarity matrix. It is recommended
when processing the ultra-dimention data.}

\item{pca.sum.prop}{A numeric indicating how many components should be reserved
in order to make this propotion of variance. The default is \code{pca.sum.prop =  0.95}.}

\item{fix.pca.comp}{A numeric variable indicating whether choosing the fixed
number of components or the fixed porpotion of variance and the default is to
choose fixed porpotion.}

\item{n.comp}{The number of components of PCA. The default is \code{n.comp = 50}.}

\item{alpha}{A power scaling for Bossa scores, representing the weight of
variable sigma value.}
}
\value{
An object including Bossa scores, Bossa disimilarity and Bossa similarity(for
\code{\link{OverlapClust}}.
}
\description{
Calculate the similarity matrix of Bossa scores which are obtained by boosting
on single attribute.
}
