% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{ClustMerge}
\alias{ClustMerge}
\title{Clust Merge}
\usage{
ClustMerge(tree.size, clu.hc, n, overlap.clu, data.simi)
}
\arguments{
\item{tree.size}{The label of each overlap clusters.}
}
\description{
With every recommended k, merge the overlap clusters depending on the distance
between those.
}
\keyword{internal}
