% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{ClustShare}
\alias{ClustShare}
\title{Calculate the share matrix among overlap clusters.}
\usage{
ClustShare(overlap.clu)
}
\arguments{
\item{overlap.clu}{The result of overlap clustering.}
}
\description{
Calculate the share matrix among overlap clusters.
}
\keyword{internal}
