% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prepare.R
\name{FindBefDe}
\alias{FindBefDe}
\title{Find DE from original overlap clusters}
\usage{
FindBefDe(mer.clu, overlap.clu, U.score.non.pca, cell.name)
}
\description{
Find DE from original overlap clusters
}
\keyword{internal}
