% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prepare.R
\name{FindHcDe}
\alias{FindHcDe}
\title{Find DE from HC with recommended k}
\usage{
FindHcDe(cell.hc.clust, U.score.non.pca, cell.name)
}
\description{
Find DE from HC with recommended k
}
\keyword{internal}
