% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prepare.R
\name{FindOverlapDe}
\alias{FindOverlapDe}
\title{Find DE from Overlap clusters}
\usage{
FindOverlapDe(clu.merge, U.score.non.pca, cell.name, n)
}
\description{
Find DE from Overlap clusters
}
\keyword{internal}
