% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{KeyFeature}
\alias{KeyFeature}
\title{Determine key features of each overlap clusters.}
\usage{
KeyFeature(cdata, cri, sum.clu)
}
\arguments{
\item{cdata}{Overlap clustering result.}

\item{cri}{A tuning parameter, if p value smaller than cri, then reject
the NULL hypothesis and merge overlap sub-clusters. And cri can be any numeric less
than \code{1}, if \code{cri = 1} then the criteria will be reset to \code{0.05/N}
(N is the number of all overlap sub-cluster), and if \code{cri = 2} then the
criteria \code{0.05/N(N-1)}.}
}
\value{
A list including the recommend parameter to cut trees in HC \code{scrit0, scrit1},
the separate overlap sub-clusters \code{sep.clu}, and highly similar sub-clusters \code{mer.clu},
statistics for merging \code{clu.math.stat}
}
\description{
Determine key features of each overlap clusters.
}
\keyword{internal}
