% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{OrderClust}
\alias{OrderClust}
\title{Reindex cluster labels in ascending order}
\usage{
OrderClust(hc, k)
}
\arguments{
\item{hc}{an object of class hclust}

\item{k}{number of cluster to be inferred from hc}
}
\description{
Given an \code{\link[stats]{hclust}} object and the number of clusters \code{k}
this function reindex the clusters inferred by \code{cutree(hc, k)[hc$order]}, so that
they appear in ascending order. This is particularly useful when plotting
heatmaps in which the clusters should be numbered from left to right.
}
\keyword{internal}
