% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clust_Method.R
\name{TestMergeClust}
\alias{TestMergeClust}
\title{Do a test before merging two overlap clusters.}
\usage{
TestMergeClust(data, r)
}
\arguments{
\item{data}{A data.frame with two columns of two overlap sub-clusters.}

\item{r}{A numeric indicating tetrachoric correlation coefficient.}
}
\description{
Do a test before merging two overlap clusters. Pairwise comparisons are processed
to show if those two clusters should be merged.
}
\keyword{internal}
