% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_method.R
\name{bossa_interactive}
\alias{bossa_interactive}
\title{Opens \code{BOSSA} results in an interactive session in a web browser.}
\usage{
bossa_interactive(object)
}
\arguments{
\item{object}{a output of \code{BossaClust} function}
}
\value{
Opens a browser window with an interactive \code{shiny} app and visualize
all pre-computed clusterings.
}
\description{
Runs interactive \code{shiny} session of \code{BOSSA} based on pre-computed clusterings.
}
\examples{
{
data(bo.simu.data)
object <- BossaClust(bo.simu.data)
if(interactive()){bossa_interactive(object)}
}
}
